requirements.txt
opencv-python
numpy
pillow
pyinstaller